pyproject.toml
[project]
name = "imgcount"
version = "0.1.0"
description = "Generate and analyze images with counted number of objects."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fal-client>=0.9.1",
    "google-genai>=1.51.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[tool.uv.workspace]
members = [
    "imgcount",
]
